+++ src/s2/s2edge_vector_shape.h
@@ -60,7 +60,7 @@ class S2EdgeVectorShape : public S2Shape {
   }
 
   // S2Shape interface:
-  int num_edges() const final { return edges_.size(); }
+  int num_edges() const final { return static_cast<int>(edges_.size()); }
   Edge edge(int e) const final {
     return Edge(edges_[e].first, edges_[e].second);
   }
@@ -68,7 +68,7 @@ class S2EdgeVectorShape : public S2Shape {
   ReferencePoint GetReferencePoint() const final {
     return ReferencePoint::Contained(false);
   }
-  int num_chains() const final { return edges_.size(); }
+  int num_chains() const final { return static_cast<int>(edges_.size()); }
   Chain chain(int i) const final { return Chain(i, 1); }
   Edge chain_edge(int i, int j) const final {
     DCHECK_EQ(j, 0);
